betterchisquare.test {betterchisquare}	R Documentation

betterchisquare.test

Description

This function performs the betterchisquare test

Usage

betterchisquare.test(x, dnull, param, dalt, lower = -Inf, upper = Inf,
  nkappa = 5)

Arguments

x
data, either a vector of values or a list with elements counts and breaks

dnull
a character string for the density under the null

dalt
a character string for the density under the alternative

lower

upper

nkappa
number of kappa values included

Examples

 betterchisquare.test(x, pnull=pnorm, palt=function(x) {pt(x, 3)})

[Package betterchisquare version 0.0.0.900 Index]